python-database-examples/requirements.txt
# 数据库学习示例 - 依赖包
# 安装命令：pip install -r requirements.txt

# MySQL驱动
mysql-connector-python>=8.0.0

# PostgreSQL驱动
psycopg2-binary>=2.9.0

# 对象存储（阿里云OSS）
oss2>=2.18.0

# 基础爬虫依赖
requests>=2.31.0
beautifulsoup4>=4.12.0

# 数据处理
pandas>=2.0.0
